Create boot image with partition table
Create a boot image with partition table. This is the sane variant for
USB flash drives and also seems to work when flashing the installer
straight onto the CM4 eMMC.

The same boot.img can still be used with rpiboot, so there seems to be
really no downside of using a partition table.

Signed-off-by: Stefan Agner <stefan@agner.ch>